fix(nhdplus_utils): join PlusFlowVAA and Elevslope tables to index of flowlines (left join) instead of inner join, so that flowlines not referenced in these tables are still included. fix(docs): some fixes to the docs so that they build; some additional cleanup of jupyter notebook output in docs and reorg

+++ sfrmaker/grid.py
@@ -309,7 +309,7 @@ class StructuredGrid(Grid):
     @property
     def isfr(self):
         return np.reshape(self.df.isfr.values,
-                          (self.nrow, self.ncol)).astype(np.int32)
+                          (self.nrow, self.ncol)).astype(int)
 
     @property
     def uniform(self):
